+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryUtil.java
@@ -339,10 +339,13 @@ public class BigQueryUtil {
   }
 
   // allowing widening narrow numeric into bignumeric
+  // allowing writing string to time
   @VisibleForTesting
   static boolean typeWriteable(LegacySQLTypeName sourceType, LegacySQLTypeName destinationType) {
     return (sourceType.equals(LegacySQLTypeName.NUMERIC)
             && destinationType.equals(LegacySQLTypeName.BIGNUMERIC))
+        || (sourceType.equals(LegacySQLTypeName.STRING)
+            && destinationType.equals(LegacySQLTypeName.TIME))
         || sourceType.equals(destinationType);
   }